requirement.txt
flask
requests
numpy
pillow